+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## Next
+
 ## 94
+
+### Features
+
+- Add `Result` schema and type to Zod library which allows for representing type with a success or error state.
 
 ### Overview